3.19，第三部分问题，被征地补贴金额公式更新

+++ langfang-shebao/src/main/java/com/ruoyi/shebao/dto/ExpropriateeSubsidyFormDto.java
@@ -99,7 +99,7 @@ public class ExpropriateeSubsidyFormDto
     private Integer flexibleEmploymentMonths;
 
     /** 已领取困难人员社会保险补贴月数 */
-    private BigDecimal difficultySubsidyMonths;
+    private Integer difficultySubsidyMonths;
 
     /** 截至基准日的年龄 */
     private Integer ageAtBaseDate;
