+++ sql/langfang-shebao.sql
@@ -139,7 +139,7 @@ create table shebao_expropriatee_subsidy
     base_date                    date                          null comment '基准日',
     employee_pension_months      int         default 0         null comment '职工身份缴纳职工养老月数',
     flexible_employment_months   int         default 0         null comment '灵活就业身份缴纳职工养老保险月数',
-    difficulty_subsidy_months    decimal(5,2) default 0.00     null comment '已领取困难人员社会保险补贴月数',
+    difficulty_subsidy_months    int         default 0         null comment '已领取困难人员社会保险补贴月数',
     age_at_base_date            int                           null comment '截至基准日的年龄',
     subsidy_years               decimal(5,2) default 0.00     null comment '被征地参保补贴年限',
     subsidy_amount              decimal(10,2) default 0.00    null comment '补贴金额',
